Big refactor: alerts
