+++ webapp/src/components/chat/ChatInput.tsx
@@ -17,7 +17,6 @@ import { useAppDispatch, useAppSelector } from '../../redux/app/hooks';
 import { RootState } from '../../redux/app/store';
 import { addAlert } from '../../redux/features/app/appSlice';
 import { editConversationInput, updateBotResponseStatus } from '../../redux/features/conversations/conversationsSlice';
-import { Alerts } from '../shared/Alerts';
 import { getErrorDetails } from '../utils/TextUtils';
 import { SpeechService } from './../../libs/services/SpeechService';
 import { updateUserIsTyping } from './../../redux/features/conversations/conversationsSlice';
@@ -215,7 +214,6 @@ export const ChatInput: React.FC<ChatInputProps> = ({ isDraggingOver, onDragLeav
             <div className={classes.typingIndicator}>
                 <ChatStatus chatState={chatState} />
             </div>
-            <Alerts />
             <div className={classes.content}>
                 <Textarea
                     title="Chat input"
