+++ webapp/src/components/chat/ChatWindow.tsx
@@ -16,7 +16,6 @@ import React from 'react';
 import { useAppSelector } from '../../redux/app/hooks';
 import { RootState } from '../../redux/app/store';
 import { FeatureKeys } from '../../redux/features/app/AppState';
-import { Alerts } from '../shared/Alerts';
 import { ChatRoom } from './ChatRoom';
 import { DeleteChatHistory } from './controls/DeleteChatHistory';
 import { ParticipantsList } from './controls/ParticipantsList';
@@ -60,11 +59,6 @@ const useClasses = makeStyles({
     controls: {
         display: 'flex',
         alignItems: 'center',
-    },
-    alerts: {
-        display: 'flex',
-        flexDirection: 'column',
-        ...shorthands.margin(0, '72px'),
     },
 });
 
@@ -172,11 +166,6 @@ export const ChatWindow: React.FC = () => {
                 />
             )}
             {selectedTab === ChatWindowTabEnum.PERSONA && <PersonaTab />}
-            {selectedTab !== ChatWindowTabEnum.CHAT && (
-                <div className={classes.alerts}>
-                    <Alerts />
-                </div>
-            )}
         </div>
     );
 };
